admin/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.27.1
google-cloud-firestore>=2.14.0
google-cloud-bigquery>=3.18.0
google-generativeai>=0.4.1
google-adk>=0.3.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.6.3
pydantic-settings>=2.2.1
pytest>=8.0.2
pytest-asyncio>=0.23.5
pytest-mock>=3.12.0